Revert "Merge branch '2.0.x' of https://github.com/MarlinFirmware/Marlin into 2.0.x"
This reverts commit 6deb7eb9f11c7007d1d1d34e07f2ef2a24b41bb4, reversing
changes made to 71a09b82efd70f70e341f0ff6ab1be174289ebe0.

+++ buildroot/share/PlatformIO/variants/FYSETC_S6/variant.h
@@ -149,12 +149,11 @@ extern "C" {
 /* HAL configuration */
 #define HSE_VALUE               12000000U
 
-#define FLASH_PAGE_SIZE         (4U * 1024U)
+#define FLASH_PAGE_SIZE			((uint32_t)(4*1024))
 
 #ifdef __cplusplus
 } // extern "C"
 #endif
-
 /*----------------------------------------------------------------------------
  *        Arduino objects - C++ only
  *----------------------------------------------------------------------------*/
